Update Index.js and TaskItem.js To Accomodate Reset Module
